Double -> int for macros int objectives + types shot -> types int

+++ lib/data/models/aliment_model.dart
@@ -15,7 +15,7 @@ class Aliment {
   String? image;
   String? unit;
   double? servingQuantity;
-  short? calories;
+  int? calories;
   double? proteins;
   double? carbohydrates;
   double? sugars;
